requirements.txt
openai
langchain
langchain-community
chromadb
tqdm
rich
fastapi
uvicorn
python-dotenv
pydantic
redis
streamlit         # (for UI)
